kira/senses/pyproject.toml
[project]
name = "kira-senses"
version = "0.1.0"
description = "Pluggable perception modules for Kira"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
vision = [
    "opencv-python>=4.8",
    "numpy>=1.24",
    "torch>=2.0",
    "transformers>=4.35",
    "pillow>=10.0",
]
hearing = [
    "numpy>=1.24",
    "torch>=2.0",
    "faster-whisper>=0.10",
    "sounddevice>=0.4",
]
voice = [
    "piper-tts>=1.2",
]
screen = [
    "mss>=9.0",
    "numpy>=1.24",
    "torch>=2.0",
    "transformers>=4.35",
    "pillow>=10.0",
]
elevenlabs = [
    "elevenlabs>=1.0",
]
all = [
    "kira-senses[vision,hearing,voice,screen]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
